RESTPLUS/requirements.txt
Flask==1.0.2
flask-restplus==0.11.0
grpcio==1.8.6
tensorflow>=1.12
numpy<=1.14.5
pillow==6.1.0
scikit-image==0.15
